Added accountsRead permission to manifest.json
The accountsRead permission is necessary because we cannot access
messenger.accounts API without it. We need access to this API to
be able to iterate through each account's mail folders to accomplish
the goal of graphing each sender in a graph to show the user where
most of their email is coming from.