Refactored main.html, Updated background.js, Added account selection
- Organized main.html into its own folder and created a separate stylesheet
and javascript file for it.
- Updated background.js with the new path to main.html.
- In the dashboard for the extension, the user can now select a specific account
from a dropdown list. When they do, it will show the name of the account as well
as how many different folders it has. Accomplishing this was my way of learning
the thunderbird accounts API and will lead into the actual functionality of
the extension later on.